MySQL linting support and fixes
allow only to save the user that has his character selected and only that character

prior was saving all user characters even when user was in the selection menu.

also there was something randomly assigning a wrong char id there maybe it would be the cause of some players could not get in.