refactor(server): deprecate old events + add new from server side
these events had to be changed to be handled from server side due to cheaters trying to spam them
you will have to update your scripts to use new. or don't.
check documentation
you will have to change some the event names others from RegisterNetEvent to AddEventHandler

only if you are using them, if you are not ignore this.